Extrais `<ListeInvitations />`
… pour alléger `<InvitationContributeur />`
